[#134766445] partial rewrite of tool buttons
- the game executable can now also be configured in the same way as the tools
- tools can have environment variables set that will be set on their process (preconfigured or custom)
- removed (pre-configured )tools can be restored
- The "primary" tool button (usually the game) can be changed (i.e. if the user always wants to start with skse)
- Visual changes (no more context menu since this was the only location we used those)

- this also contains a fix to the state-proxy so it supports delete operations

+++ src/extensions/gamemode_management/reducers/settings.ts
@@ -18,18 +18,22 @@ export const settingsReducer: IReducerSpec = {
                      ['discovered', payload.gameId, 'tools', payload.toolId],
                      payload.result);
     },
-    [actions.removeDiscoveredTool]: (state, payload) => {
-      // custom added tools can be deleted. pre-configured ones have static discovery data
-      // in knownTools so they would always reappear. Therefore we just set them to hidden
-      if (getSafe(state,
+    [actions.setToolVisible]: (state, payload) => {
+      // custom added tools can be deleted so we do that instead of hiding them
+      if (!payload.visible && getSafe(state,
                   ['discovered', payload.gameId, 'tools', payload.toolId, 'custom'],
                   false)) {
         return deleteOrNop(state, ['discovered', payload.gameId, 'tools', payload.toolId]);
       } else {
         return setSafe(state,
                        ['discovered', payload.gameId, 'tools', payload.toolId, 'hidden'],
-                       true);
+                       !payload.visible);
       }
+    },
+    [actions.setGameParameters]: (state, payload) => {
+      // this is effectively the same as addDiscoveredGame but if we were adding safety
+      // checks this would have other checks than addDiscoveredGame
+      return merge(state, ['discovered', payload.gameId], payload.parameters);
     },
     [actions.setGameHidden]: (state, payload) => {
       return setSafe(state, ['discovered', payload.gameId, 'hidden'], payload.hidden);
